nv04-nv40: avoid some METHOD_CNT insanity with GDI_RECTANGLE_TEXT..
For some reason any activity on another subchannel between setting rect
colour, and emitting the rectangle will cause a METHOD_CNT exception.

If the pushbuf gets flushed between these the drm will insert commands
for the fence counter, and expose this odd aspect of the GDI_RECT object.

We can improve how we do this, and I probably will at some point.. But
getting a workable mm is a higher priority :)