nv50/exa: make sure tesla_scratch usage is tracked
Because we were emitting this once during init, and never again, the
bufmgr thought the buffer was unused and kicked it out in low memory
situtations.  This caused everything we keep there to get corrupted,
and made the engine very unhappy!

This commit causes PrepareComposite() to be repeated after a flush,
not sure how much of an impact this'll have (hopefully minimal).
After everything's working nicely this may be reworked to have a
less severe reemit.